+++ CHANGELOG.md
@@ -4,4 +4,8 @@
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- 암호화/복호화 기능 (AES-128/192/256 CBC/GCM, DES CBC, 3DES CBC)
+- PBKDF2WithHmacSHA256 기반 패스워드 키 파생
+- 해시 생성 기능 (MD5, SHA-1, SHA-256, SHA-512)
+- Base64 인코딩/디코딩 기능
+- Crypto Tool Window (암호화/복호화, 해시, Base64 탭 구성)